web/requirements.txt
appnope==0.1.0
decorator==4.1.2
Django==1.11.6
django-admin-honeypot==1.0.0
django-environ==0.4.4
django-extensions==1.9.6
django-flat-responsive==2.0
django-redis==4.8.0
django-render-block==0.5
django-solo==1.1.2
django-split-settings==0.2.5
django-templated-email==2.2.0
django-wysiwyg-redactor==0.5.1
gunicorn==19.7.1
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.2
redis==2.10.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
